requirements.txt
textual
rich
i2plib
pyperclip
PyQt6
qasync
Pillow
